Better token handling
This PR adds better token handling on the server-side. It removes the backend API token from the `session`, since we do all calls on the server-side now, which is much cleaner. It also syncs the expiry of frontend and backend token (this fixes #467) and moves the locale to the user session.